# packages/api-client/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deskflow API",
    "description": "Flex workspace booking. See docs/TECHNICAL_DESIGN.md.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/sites/{site_id}/floors": {
      "get": {
        "tags": [
          "spaces"
        ],
        "summary": "List Floors",
        "operationId": "list_floors_v1_sites__site_id__floors_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FloorOut"
                  },
                  "title": "Response List Floors V1 Sites  Site Id  Floors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/sites": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Site",
        "description": "FR-8.1. The timezone is the site's, and it is the authority for \"a day\" (TDD \u00a75).",
        "operationId": "create_site_v1_admin_sites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__v1__admin__SiteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/sites/{site_id}/floors": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Floor",
        "operationId": "create_floor_v1_admin_sites__site_id__floors_post",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloorIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Floors",
        "operationId": "list_floors_v1_admin_sites__site_id__floors_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FloorSummaryOut"
                  },
                  "title": "Response List Floors V1 Admin Sites  Site Id  Floors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/floors/{floor_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Open Floor",
        "operationId": "open_floor_v1_admin_floors__floor_id__get",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorEditorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/floors/{floor_id}/layout": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Save Layout",
        "description": "Explicit, batched save (TDD \u00a714.3). Writes a draft; employees see nothing yet.",
        "operationId": "save_layout_v1_admin_floors__floor_id__layout_put",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloorLayout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutSavedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/floors/{floor_id}/draft": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Discard Draft",
        "operationId": "discard_draft_v1_admin_floors__floor_id__draft_delete",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/floors/{floor_id}/plan": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Upload Plan",
        "description": "Upload an image or PDF; the server rasterizes and returns dimensions (TDD \u00a714.3).\n\nThe asset is attached to the DRAFT, not to the live floor. Replacing a plan is an\nedit like any other and must not become visible to employees before publish \u2014 and\nthe aspect-ratio warning (TDD \u00a714.2) has to be reviewable before it takes effect.",
        "operationId": "upload_plan_v1_admin_floors__floor_id__plan_post",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_plan_v1_admin_floors__floor_id__plan_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/floors/{floor_id}/publish": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Preflight",
        "description": "What publishing would do, including the bookings it would strand (TDD \u00a714.3).",
        "operationId": "preflight_v1_admin_floors__floor_id__publish_get",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Preflight V1 Admin Floors  Floor Id  Publish Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Publish",
        "operationId": "publish_v1_admin_floors__floor_id__publish_post",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Publish V1 Admin Floors  Floor Id  Publish Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/floors/{floor_id}/codes": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Taken Codes",
        "description": "Codes already used elsewhere on this site.\n\nBulk creation generates names from a pattern (`4F-A-{01..24}`), and the constraint\nit can violate is site-wide, not floor-wide. Handing the editor the taken codes lets\nit show the clash while the admin is still typing the pattern, instead of failing\nthe whole batch on save.",
        "operationId": "taken_codes_v1_admin_floors__floor_id__codes_get",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Floor Id"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prefix"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Taken Codes V1 Admin Floors  Floor Id  Codes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FloorEditorOut": {
        "properties": {
          "floor": {
            "$ref": "#/components/schemas/FloorSummaryOut"
          },
          "site": {
            "$ref": "#/components/schemas/app__api__v1__admin__SiteOut"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "layout": {
            "$ref": "#/components/schemas/FloorLayout"
          },
          "is_draft": {
            "type": "boolean",
            "title": "Is Draft"
          },
          "draft_updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Updated At"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupOut"
            },
            "type": "array",
            "title": "Groups"
          }
        },
        "type": "object",
        "required": [
          "floor",
          "site",
          "plan",
          "layout",
          "is_draft",
          "draft_updated_at",
          "groups"
        ],
        "title": "FloorEditorOut",
        "description": "Everything the editor needs to open a floor, in one round trip."
      },
      "FloorIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FloorIn"
      },
...
      "FloorOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "site_id": {
            "type": "string",
            "format": "uuid",
            "title": "Site Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          },
          "plan_width_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Width Px"
          },
          "plan_height_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Height Px"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "name",
          "ordinal",
          "plan_width_px",
          "plan_height_px"
        ],
        "title": "FloorOut"
      },
      "FloorSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "site_id": {
            "type": "string",
            "format": "uuid",
            "title": "Site Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          },
          "plan_width_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Width Px"
          },
          "plan_height_px": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Height Px"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "resource_count": {
            "type": "integer",
            "title": "Resource Count"
          },
          "has_draft": {
            "type": "boolean",
            "title": "Has Draft"
          }
        },
        "type": "object",
        "required": [
          "id",
          "site_id",
          "name",
          "ordinal",
          "plan_width_px",
          "plan_height_px",
          "published_at",
          "resource_count",
          "has_draft"
        ],
        "title": "FloorSummaryOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LayoutSavedOut": {
        "properties": {
          "base_version": {
            "type": "integer",
            "title": "Base Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "layout": {
            "$ref": "#/components/schemas/FloorLayout"
          }
        },
        "type": "object",
        "required": [
          "base_version",
          "updated_at",
          "layout"
        ],
        "title": "LayoutSavedOut"
      },
...
      "SiteIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "opening_hours": {
            "additionalProperties": true,
            "type": "object",
            "title": "Opening Hours"
          }
        },
        "type": "object",
        "required": [
          "name",
          "timezone"
        ],
        "title": "SiteIn"
      },
...
      "app__api__v1__admin__SiteOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "timezone",
          "address"
        ],
        "title": "SiteOut"
      },
...
    }
  }
}